requirements.txt
streamlit==1.28.1
Pillow==10.1.0
replicate==0.15.4
python-dotenv==1.0.0
